requirements.txt
pymongo
python-dotenv
passlib[bcrypt]
customtkinter
dnspython
